+++ src/components/index.ts
@@ -1,3 +1,4 @@
+export { Dialog } from './dialog';
 export { Search } from './search';
 export { StockCard } from './stock-card';
 export { StockChart } from './stock-chart';